% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCleaning.R
\name{CleanDataByFylke}
\alias{CleanDataByFylke}
\title{CleanDataByFylke}
\usage{
CleanDataByFylke(d, FylkeData, myfylke)
}
\arguments{
\item{d}{a}

\item{FylkeData}{a}

\item{myfylke}{a}
}
\description{
CleanDataByFylke
}
