% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePlots.R
\name{CreatePlots2}
\alias{CreatePlots2}
\title{CreatePlots2}
\usage{
CreatePlots2(d1, weeknow, Ukenummer, Fylkename, S, mytittle)
}
\arguments{
\item{d1}{a}

\item{weeknow}{a}

\item{Ukenummer}{a}

\item{Fylkename}{a}

\item{S}{a}

\item{mytittle}{a}
}
\description{
CreatePlots2
}
