% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePlots.R
\name{CreatePlotsNorway}
\alias{CreatePlotsNorway}
\title{CreatePlots1}
\usage{
CreatePlotsNorway(d, weeknow, Ukenummer, title, yrange)
}
\arguments{
\item{d}{a}

\item{weeknow}{a}

\item{Ukenummer}{a}

\item{title}{a}

\item{yrange}{a}
}
\description{
CreatePlots1
}
