% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCleaning.R
\name{DeleteOldDatasets}
\alias{DeleteOldDatasets}
\title{DeleteOldDatasets}
\usage{
DeleteOldDatasets(raw = list.files(fhi::DashboardFolder("data_raw"),
  "^partially_formatted_"),
  dat_in = list.files(fhi::DashboardFolder("data_raw"), "in_default"),
  clean = list.files(fhi::DashboardFolder("data_clean"), "done_"))
}
\arguments{
\item{raw}{a}

\item{dat_in}{a}

\item{clean}{a}
}
\description{
DeleteOldDatasets
}
