% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_CreatePlots.R
\name{ggplot_CreatePlots1}
\alias{ggplot_CreatePlots1}
\title{CreatePlots1}
\usage{
ggplot_CreatePlots1(d, weeknow, Ukenummer, title, yrange)
}
\arguments{
\item{d}{a}

\item{weeknow}{a}

\item{Ukenummer}{a}

\item{title}{a}

\item{yrange}{a}
}
\description{
CreatePlots1
}
