% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_CreatePlots.R
\name{ggplot_CreatePlotsNorway}
\alias{ggplot_CreatePlotsNorway}
\title{CreatePlots1}
\usage{
ggplot_CreatePlotsNorway(data_long, weeknow, Ukenummer, title, yrange)
}
\arguments{
\item{data_long}{a}

\item{weeknow}{a}

\item{Ukenummer}{a}

\item{title}{a}

\item{yrange}{a}
}
\description{
CreatePlots1
}
