% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_CreatePlots.R
\name{ggplot_CreatePlotsNorwayByAge}
\alias{ggplot_CreatePlotsNorwayByAge}
\title{CreatePlotsNorwayByAge}
\usage{
ggplot_CreatePlotsNorwayByAge(data_long, weeknow, Ukenummer, Fylkename, S,
  mytittle)
}
\arguments{
\item{data_long}{a}

\item{weeknow}{a}

\item{Ukenummer}{a}

\item{Fylkename}{a}

\item{S}{a}

\item{mytittle}{a}
}
\description{
CreatePlotsNorwayByAge
}
