% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_CreatePlots.R
\name{ggplot_plot_ages}
\alias{ggplot_plot_ages}
\title{ggplot_plot_1}
\usage{
ggplot_plot_ages(location_code, x_tag)
}
\arguments{
\item{location_code}{a}

\item{x_tag}{a}
}
\description{
ggplot_plot_1
}
