% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_CreatePlots.R
\name{ggplot_plot_total}
\alias{ggplot_plot_total}
\title{ggplot_plot_total}
\usage{
ggplot_plot_total(location_code, x_tag)
}
\arguments{
\item{location_code}{a}

\item{x_tag}{a}
}
\description{
ggplot_plot_total
}
