% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCleaning.R
\name{IdentifyInOutDoc}
\alias{IdentifyInOutDoc}
\title{IdentifyInOutDoc}
\usage{
IdentifyInOutDoc(raw = list.files(fhi::DashboardFolder("data_raw"),
  "^partially_formatted_"), fylke = fread(system.file("extdata",
  "fylke.csv", package = "sykdomspulspdf")),
  indoc = list.files(fhi::DashboardFolder("data_raw"), "in_"))
}
\arguments{
\item{raw}{a}

\item{fylke}{a}

\item{indoc}{a}
}
\description{
IdentifyInOutDoc
}
