% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemplatesCopyRemove.R
\name{sykdompulspdf_template_copy}
\alias{sykdompulspdf_template_copy}
\title{hi}
\usage{
sykdompulspdf_template_copy(output_dir, SYNDROM)
}
\arguments{
\item{output_dir}{a}

\item{SYNDROM}{a}
}
\description{
hi
}
